move cli utils in cli directory
