+++ src/cli/mod.rs
@@ -2,3 +2,5 @@ pub mod cardano;
 pub use cardano::*;
 pub mod cli;
 pub use cli::*;
+pub mod utils;
+pub use utils::*;
