+++ src/cli/utils/terminal.rs
