+++ src/main.rs
@@ -2,7 +2,6 @@ use cli::Cli;
 use structopt::StructOpt;
 mod cli;
 mod tests;
-mod utils;
 
 #[tokio::main]
 async fn main() {
